chore: Add different Sample Projects
BLub
