+++ .run/MainKt.run.xml
@@ -1,11 +1,11 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="MainKt" type="JetRunConfigurationType" nameIsGenerated="true">
-    <option name="MAIN_CLASS_NAME" value="com.example.samples.MainKt" />
-    <module name="vehicle-app-java-sdk.samples.main" />
+    <option name="MAIN_CLASS_NAME" value="com.example.java.app.MainKt" />
+    <module name="vehicle-app-java-sdk.samples.java-app.main" />
     <shortenClasspath name="NONE" />
     <method v="2">
       <option name="Make" enabled="true" />
-      <option name="Gradle.BeforeRunTask" enabled="true" tasks="assembleDist" externalProjectPath="$PROJECT_DIR$/samples" vmOptions="" scriptParameters="" />
+      <option name="Gradle.BeforeRunTask" enabled="true" tasks="installDist" externalProjectPath="$PROJECT_DIR$/samples/java-app" vmOptions="" scriptParameters="" />
     </method>
   </configuration>
 </component>
