+++ gradle/libs.versions.toml
@@ -1,9 +1,8 @@
 [versions]
 agp = "8.5.2"
 detektFormatting = "1.23.6"
-kotlin = "2.0.0"
+kotlin = "1.9.22"
 junitJupiter = "5.10.1"
-kotestRunnerJunit5 = "5.9.1"
 detekt = "1.23.6"
 coroutines = "1.8.1"
 kotest = "5.7.2"
@@ -15,7 +14,6 @@ kuksa-java-sdk = "0.3.1"
 [libraries]
 junit-jupiter-api = { module = "org.junit.jupiter:junit-jupiter-api", version.ref = "junitJupiter" }
 junit-jupiter-engine = { module = "org.junit.jupiter:junit-jupiter-engine", version.ref = "junitJupiter" }
-kotest-runner-junit5 = { module = "io.kotest:kotest-runner-junit5", version.ref = "kotestRunnerJunit5" }
 detekt-formatting = { module = "io.gitlab.arturbosch.detekt:detekt-formatting", version.ref = "detektFormatting" }
 kotlinx-coroutines-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "coroutines" }
 kotest = { group = "io.kotest", name = "kotest-runner-junit5", version.ref = "kotest" }
