+++ samples/android-app/.gitignore
