+++ samples/android-app/build.gradle.kts
@@ -1,0 +1,37 @@
+plugins {
+    id("com.android.application")
+    id("org.jetbrains.kotlin.android")
+    id("org.eclipse.velocitas.vss-processor-plugin") // version <VERSION>
+}
+
+vssProcessor {
+    searchPath = "vss"
+}
+
+android {
+    namespace = "com.example.android.app"
+    compileSdk = 34
+
+    defaultConfig {
+        minSdk = 24
+        targetSdk = 34
+    }
+    packaging {
+        resources.pickFirsts.add("**")
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
+    }
+    kotlinOptions {
+        jvmTarget = "11"
+    }
+}
+
+dependencies {
+    implementation(project(":sdk"))
+
+    implementation("org.eclipse.kuksa:vss-core:0.3.1")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.8.1")
+}
