+++ samples/android-app/vss/vss_rel_4.0.yaml
